block: elevator: cfq is used by default when the governor is not found
Some init scripts in the device tree will be changed to bfq as the main I/O governor. Since it is not stable on the existing kernel, we have removed it. In order to avoid accidents, the speed will be written from the user space. When the governor cannot be found after the controller, we use cfq as the default governor by default.